requirements.txt
numpy
opencv-python
PyQt5
PyYAML
torch
ultralytics
easyocr
